separated logic.js into function file and buttonClick file
